folder view and deletion fixes

+++ src/routes/delete.js
@@ -1,19 +1,30 @@
 import dotenv from 'dotenv';
 dotenv.config();
-import { DeleteObjectCommand } from "@aws-sdk/client-s3";
+
 import { getFileById, deleteFileById } from '../services/fileService.js';
 import { checkAuthenticated } from './auth.js';
 import express from 'express';
 const router = express.Router();
 import cookieParser from 'cookie-parser';
-import { b2Client } from '../config/backblaze.js';
 
 import { deleteFolderAndContents } from '../models/folderModel.js';
+import { getFolderById } from '../services/folderService.js';
 
 router.use(cookieParser(process.env.SESSION_SECRET));
 
 router.delete('/folder/:id', checkAuthenticated, async (req, res) => {
     const folderId = req.params.id;
+
+    const folder = await getFolderById(folderId);
+    if (!folder) {
+        return res.status(404).json({ error: 'Folder not found' });
+    }
+    const currentUser = req.signedCookies.dbUser;
+    if (currentUser.id !== folder.userid) {
+        return res.status(403).json({ error: 'Not authorized to delete this folder' });
+    }
+    
+    
     try {
         await deleteFolderAndContents(folderId);
         return res.status(200).send('Deleted folder: ' + folderId);
@@ -38,19 +49,6 @@ router.delete('/file/:id', checkAuthenticated, async (req, res) => {
     const currentUser = req.signedCookies.dbUser;
     if (currentUser.id !== file.userid) {
         return res.status(403).json({ error: 'Not authorized to delete this file' });
-    }
-
-    // Delete the file from S3
-    const params = {
-        Bucket: process.env.S3_BUCKET_NAME,
-        Key: fileId
-    };
-    const command = new DeleteObjectCommand(params);
-    try {
-        await b2Client.send(command);
-    } catch (error) {
-        console.error('Error deleting file from S3', error);
-        return res.status(500).send('Error deleting file from S3');
     }
 
     // Delete the file record from the database
